feat: add --wait flag for devspace logs
